requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
PyYAML
Flask-Cors
